Add 'see all' link to <Category>
Add link to upper right corner of <Category>, imitating Trip Advisor.

Bug: When a browser window is less than its max size on a laptop,
the link sits above empty space. It should be above the upper right
corner of the last <Event> in a row of <Event>s within the <Category>.

This problem doesn't exist if the browser window is maximized. The link
is right above the upper right corner of the fourth and final <Event>.